9/12

+++ Assets/Scripts/Utils/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(int damage,GameObject attacker);
+    void TakeDamage(float damage,GameObject attacker);
 }
